+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change Log
+## [5.49.0](https://github.com/plivo/plivo-dotnet/tree/v5.49.0) (2024-09-30)
+**Feature - Adding new param support for Number Masking session with single party **
+- Added `create_session_with_single_party`, `virtual_number_cooloff_period` and `force_pin_authentication` attributes in Masking Session
+
 ## [5.48.1](https://github.com/plivo/plivo-dotnet/tree/v5.48.1) (2024-09-06)
 **Feature - Adding more attribute on mdr object**
 - Added `message_sent_time`, `message_updated_time` and `error-message` on get and list Message API